handlers/handlers/utils/utils/requirements.txt
python-telegram-bot==20.0
psycopg2-binary==2.9.3
stripe==2.56.0
i18n==0.10.1
python-dotenv==0.19.2